clean up filenames
